requirements.txt
python-dotenv
py-cord==2.0.0b7
PyNaCl